fixup generated api
make it work such that we have a type per CTIDH key size.
